Group fsync operations in WAL for better throughput (#211)
Instead of fsyncing every single batch, the wal runs a go-routine,
grouping multiple batches together.

Added 2 new methods in `WAL` interface:
 * `AppendAsync()` writes to wal, no sync, entries are not visibile
* `Sync()` triggers a background flush (unless there's already a pending
one) and waits for its completion

Note: this is no change yet in the `follower_controller` to take
advantage of this. I'll do a refactor later to move all the insertion
into DB in a separate background task.