chaos-mesh (#261)
This was a challenge to get working!

This builds on the docker build step to also run chaos-mesh. This first
iteration spins up an oxia cluster and a perf client. The chaos-mesh
workflow is configured with a single entry that will kill one of the
server pods.

There are currently no formal assertions yet, so as long as the full
process runs, the action will pass. Verification can be performed
manually by reviewing the coordinator & perf logs which should
illustrate pod recovery and uninterrupted perf execution.