After coordinator restarts, reassess the state of the entire ensemble (#321)
If the coordinator restarts and a follower from a previous election was
not properly fenced before, it would currently continue in the same
state, as the new coordinator instance will not attempt to fix it.

When coordinator restarts, it needs to check the state of every node in
the ensemble and not just the leader. If there are followers that are
not in the latest term, we will trigger a new election.