Added readiness support in server health-check (#351)
The oxia server should not considered to be "ready" to be put in the
load-balancer rotation for service discovery until it has received the
shards mapping from the coordinator. This can happen 10s of seconds
after restarting a storage pod.

Added gRPC service `oxia-readiness` that can be probed from K8S
readiness probe, so that we don't get a client to ask for service
discovery to a pod that does not have that information yet.