Fixed node controller assignemnts update after pods restart (#368)
If a storage node restarts and its IP stop responding, the coordinator
health check will detect it in a ~2seconds period, though the component
that sends the cluster assignements updates will not detect the issue.

One of the reason is that it's using a grpc stream, so there is no
timeout on the send operation (because it would apply to the lifetime of
the stream itself, which instead needs to be long-lived).

The storage servers need to receive the new cluster assignments after
restart, otherwise they cannot pass it back to their clients.

To fix the issue, we close the assignments stream whenever the
health-check failure is triggered. This will ensure that coordinator
will retry after the restart.