fix(coordinator): panic when BadVersion in coordinator (#512)
## Motivation

The coordinator will stop all the operations when it stuck in the
metadata `bad version` case. this PR is trying to trigger panic when bad
version to help auto-restart. the bad version might caused by many
cases:


1. operator/maintainer update the status config.
2. another coordinator updated metadata before shutdown.


> the coordinator is using lock to ensure all the metadata operation is
invoke sequentially. which means we don't have race conditions in the
process.

## Modification

- panic directly when met `BadVersion`

## Alternatives

- Update to the latest version when update metadata get `BadVersion`.
but we have to change some logics to ensure get version before building
the new status to avoid unexpected override.