Add term check on the internal streams, to avoid wrong closures of leader controller (#513)
During a leader election, an old leader can force back the new leader
into a follower mode.

This does not affect the correctness (eg: the old leader's requests are
then ignored because the term is invalid), though it causes the shard to
be unavailable until a new leader election is triggered.

The simplified sequence is like:
 1. `oxia-1` is the old leader in term: 4
 2. Coordinator initiates a new leader election and fences all the nodes
3. `oxia-0` and `oxia-2` are fenced correctly, while oxia-1 is ignored,
and will be retried in background
 4. `oxia-2` is elected leader (term: 5)
5. `oxia-1` is still running the leader controller (because it never got
successfully fenced), and it keeps trying to replicated to `oxia-0` and
`oxia-2` .


The replicate requests keep failing (because the term has already
changed to 5), though the first request has the effect of putting oxia-2
back in follower mode.

We are now in a state where coordinator is happy thinking `oxia-2` as
the leader, though `oxia-2` went back to follower mode. A new leader
election is not triggered because `oxia-2` looks healthy overall (the
health check is done at the pod level, not at shard level).