fix(comparer): disable slash abbreviated key to avoid unexpected behaviour (#684)
## Motivation

There's the root cause of the issue mentioned in #681. The index batch
skiplist will use the abbreviatedKey to do the first comparison.


https://github.com/cockroachdb/pebble/blob/939f3193b60696df444b3c76e07f0020954ea87c/internal/batchskl/iterator.go#L105-L116

Our configuration is default, so that will cause an unexpected result
because `/` is less than `a` lexicographically.

Therefore, I want to find a way to implement our self-abbreviated key
method. but we need to get a number of `/` on the origin key. That seems
a bit expensive for us. I choose to simply disable slash abbreviatedKey
and then fall back to comparison directly. Maybe we can consider more to
create a function that can support that in the future.

## Modification

- Disable slash `abbreviatedKey` to give them the same value.