Simplify the merging of GetResponse results from muiltple shards (#705)
## Motivation

A multi-shard get will return at most one of the returned responses.
That correct response can be selected without collecting and sorting the
results, allowing us to save both memory and compute.

## Modification

Keep track of the `selected` response. In the get completion callback,
decide if the new response is better.