Fix WriteStream regression: don't send an error client close
The refactoring in c90c79172d6cda7ae9ff199a8a9bf8a275fd5d61 makes
a change that sends an unknown error message to the client when it
encounters either a read EOF or an internal context is closed.  This
can confuse a client that doesn't expect to see any more messages
after closing the request side of the stream.

This change suppresses the message upon EOF.